% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnBirdWDI.R
\name{WorldBankData_function}
\alias{WorldBankData_function}
\title{A Global Threatened Birds Mapping Function}
\usage{
WorldBankData_function(ans = TRUE)
}
\arguments{
\item{Yes}{Do you want to map bitds? Defaults to TRUE.}
}
\description{
This function allows you to visualize global threatened birds distribution
}
\examples{
EnBirdWDI_function()
}
\keyword{Birds}
